Added parseVersionString to Utils| moved Utils to shared dir| type fixes
